SW-7755 Remove voting context/provider
Replace VotingProvider/VotingContext with a useVotingData hook. RTK Query's
cache makes the fetch-once provider redundant: each consumer calls the hook
directly and shares the same deduped getProjectVotes request. The invalid-
project redirect and fetch-error toast move into VotingWrapper, the shared
layout for the voting pages.
